Split up classic/CPE raw packet processing into separate ClassicProtocol.cs
