+++ MCGalaxy/MCGalaxy_.csproj
@@ -599,6 +599,7 @@
     <Compile Include="Modules\Relay\RelayBotSender.cs" />
     <Compile Include="Modules\Security\IPThrottler.cs" />
     <Compile Include="Network\BaseWebSocket.cs" />
+    <Compile Include="Network\ClassicProtocol.cs" />
     <Compile Include="Network\CPESupport.cs" />
     <Compile Include="Network\Heartbeat\ClassiCube.cs" />
     <Compile Include="Network\Heartbeat\Heartbeat.cs" />
@@ -727,32 +728,7 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Commands\Maintenance" />
-    <Folder Include="Config\Permissions" />
-    <Folder Include="Database\Stats" />
-    <Folder Include="Games\Weapons" />
-    <Folder Include="Games\RoundsGame" />
-    <Folder Include="Generator\LibNoise" />
-    <Folder Include="Generator\Classic" />
-    <Folder Include="Generator\Realistic" />
-    <Folder Include="Modules\Awards" />
-    <Folder Include="Modules\Moderation" />
-    <Folder Include="Modules\Moderation\Notes" />
-    <Folder Include="Modules\Security" />
-    <Folder Include="Modules\Relay\Discord" />
-    <Folder Include="Modules\Relay" />
-    <Folder Include="Network\Heartbeat" />
-    <Folder Include="Modules\Relay\IRC" />
-    <Folder Include="Network\Utils" />
-    <Folder Include="Network\Packets" />
-    <Folder Include="Entity" />
-    <Folder Include="Modules" />
-    <Folder Include="Server\Maintenance" />
-    <Folder Include="Server\Authentication" />
-    <Folder Include="util\Formatting" />
-    <Folder Include="util\Threading" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
