+++ MCGalaxy/Network/Player.Networking.cs
@@ -25,93 +25,8 @@ using BlockRaw = System.Byte;
 
 namespace MCGalaxy 
 {
-    public partial class Player : IDisposable, INetProtocol 
+    public partial class Player : IDisposable
     {
-        int INetProtocol.ProcessReceived(byte[] buffer, int bufferLen) {
-            int read = 0;
-            try {
-                while (read < bufferLen) {
-                    int packetLen = HandlePacket(buffer, read, bufferLen - read);
-                    // Partial packet received
-                    if (packetLen == 0) break;
-                    
-                    // Client was forced disconnected
-                    if (packetLen == -1) return bufferLen;
-                    
-                    // Packet processed, onto next
-                    read += packetLen;
-                }
-            } catch (Exception ex) {
-                Logger.LogError(ex);
-            }
-            return read;
-        }
-
-        int HandlePacket(byte[] buffer, int offset, int left) {
-            switch (buffer[offset]) {
-                case Opcode.Ping:              return 1;
-                case Opcode.Handshake:         return HandleLogin(buffer, offset, left);
-                case Opcode.SetBlockClient:    return HandleBlockchange(buffer, offset, left);
-                case Opcode.EntityTeleport:    return HandleMovement(buffer, offset, left);
-                case Opcode.Message:           return HandleChat(buffer, offset, left);
-                case Opcode.CpeExtInfo:        return HandleExtInfo(buffer, offset, left);
-                case Opcode.CpeExtEntry:       return HandleExtEntry(buffer, offset, left);
-                case Opcode.CpePlayerClick:    return HandlePlayerClicked(buffer, offset, left);
-                case Opcode.CpeTwoWayPing:     return HandleTwoWayPing(buffer, offset, left);   
-                case Opcode.CpePluginMessage:  return HandlePluginMessage(buffer, offset, left);   
-                
-                case Opcode.CpeCustomBlockSupportLevel:
-                    return left < 2 ? 0 : 2; // only ever one level anyways
-                default:
-                    Leave("Unhandled opcode \"" + buffer[offset] + "\"!", true);
-                    return -1;
-            }
-        }
-        
-        #if TEN_BIT_BLOCKS
-        BlockID ReadBlock(byte[] buffer, int offset) {
-            BlockID block;
-            if (hasExtBlocks) {
-                block = NetUtils.ReadU16(buffer, offset);
-            } else {
-                block = buffer[offset];
-            }
-            
-            if (block > Block.MaxRaw) block = Block.MaxRaw;
-            return Block.FromRaw(block);
-        }
-        #else
-        BlockID ReadBlock(byte[] buffer, int offset) { return Block.FromRaw(buffer[offset]); }
-        #endif
-                
-        int HandleBlockchange(byte[] buffer, int offset, int left) {
-            int size = 1 + 6 + 1 + (hasExtBlocks ? 2 : 1);
-            if (left < size) return 0;
-            if (!loggedIn)   return size;
-            
-            ProcessBlockchange(buffer, offset);
-            return size;
-        }
-        
-        int HandleMovement(byte[] buffer, int offset, int left) {
-            int size = 1 + 6 + 2 + (hasExtPositions ? 6 : 0) + (hasExtBlocks ? 2 : 1);
-            if (left < size) return 0;
-            if (!loggedIn)   return size;
-            
-            ProcessMovement(buffer, offset);
-            return size;
-        }
-        
-        int HandleChat(byte[] buffer, int offset, int left) {
-            const int size = 1 + 1 + 64;
-            if (left < size) return 0;
-            if (!loggedIn)   return size;
-            
-            ProcessChat(buffer, offset);
-            return size;
-        }
-
-        
         public void Send(byte[] buffer)  { Socket.Send(buffer, SendFlags.None); }
         public void Send(byte[] buffer, bool sync) {
             Socket.Send(buffer, sync ? SendFlags.Synchronous : SendFlags.None);
